refactor: move storage columns
