+++ src/containers/Storage/StorageGroups/columns/StorageGroupsColumns.scss
@@ -1,4 +1,4 @@
-.global-storage-groups {
+.ydb-storage-groups-columns {
     &__vdisks-column,
     &__disks-column {
         overflow: visible; // to enable stacked disks overflow the row
